Batch decorator
